requirements.txt
streamlit>=1.28.0
openai>=1.3.0
pillow>=10.0.0
python-dotenv>=1.0.0
concurrent-futures>=3.1.1